changing verif isAdmin on commentController

+++ src/Controllers/CommentController.php
@@ -6,11 +6,15 @@ use App\Repositories\ArticleRepository;
 use App\Repositories\CommentRepository;
 use App\Repositories\UserRepository;
 use App\Services\AuthService;
+use App\Services\UtilService;
 
 class CommentController extends AbstractController
 {
+
     /**
      * Validate the comment from article page
+     * @param $id
+     * @return void
      */
     public function validate($id)
     {
@@ -18,22 +22,45 @@ class CommentController extends AbstractController
 
         $commentRepository = new CommentRepository();
         $comment = $commentRepository->findOne($id);
+
+        if($comment->getId() === null) {
+            header("location: /");
+            return;
+        }
+
+        if(!AuthService::isAdmin()){
+            header("location: /");
+            return;
+        }
+
         $comment->setPublished(true);
         $commentRepository->save($comment);
 
         \header("location: /article/{$comment->getArticleId()}");
         return;
     }
+
 
     /**
      * Validate the comment from comments page
+     * @param $id
+     * @return void
      */
     public function validateFromComments($id)
     {
-    AuthService::checkAdmin(pathToRedirect: "/articles");
-
         $commentRepository = new CommentRepository();
         $comment = $commentRepository->findOne($id);
+
+        if($comment->getId() === null) {
+            header("location: /");
+            return;
+        }
+
+        if(!AuthService::isAdmin()){
+            header("location: /");
+            return;
+        }
+
         $comment->setPublished(true);
         $commentRepository->save($comment);
 
@@ -41,37 +68,64 @@ class CommentController extends AbstractController
         return;
     }
 
+
     /**
      * Delete comment from the article page
+     * @param $id
+     * @return void
      */
-    public function delete($id) // OK
+    public function delete($id)
     {
-//        pathToRedirect: "/articles"
-    AuthService::checkAdmin();
-
         $commentRepository = new CommentRepository();
         $comment = $commentRepository->findOne($id);
+
+        if($comment->getId() === null) {
+            header("location: /articles");
+            return;
+        }
+
+        if(!AuthService::isAdmin()){
+            header("location: /articles");
+            return;
+        }
+
         $commentRepository->deleteComment($id);
 
         \header("location: /article/{$comment->getArticleId()}");
         return;
     }
 
+
     /**
      * Delete comment from comments page
+     * @param $id
+     * @return void
      */
     public function deleteFromComments($id) // OK
     {
-    AuthService::checkAdmin(pathToRedirect: "/articles");
         $commentRepository = new CommentRepository();
+        $comment = $commentRepository->findOne($id);
+
+        if($comment->getId() === null) {
+            header("location: /articles");
+            return;
+        }
+
+        if(!AuthService::isAdmin()){
+            header("location: /articles");
+            return;
+        }
+
         $commentRepository->deleteComment($id);
 
         \header("location: /commentaires");
         return;
     }
 
+
     /**
      * Display comments
+     * @return null
      */
     public function checkAllComments()
     {
